定时任务预警预警 规则配置页面相关功能 导出Excel功能

+++ src/main/java/com/fq/modules/was/web/entity/setting/SysConfig.java
@@ -33,9 +33,9 @@ public class SysConfig{
 	 */
 	private Integer wasNum;
 	/**
-	 * 小数 例如0.31
+	 * 小数 例如31
 	 */
-	private BigDecimal wasPercent;
+	private Integer wasPercent;
 	/**
 	 * 来源系统，比如Pizza（P
 	 */
